backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.2
python-dotenv==1.0.0
email-validator==2.1.0